[#69]refactor: Main 페이지에서 컴포넌트들 분리
